requirements.txt
decorator==4.4.2
grpcio==1.28.1
grpcio-tools==1.28.1
networkx==2.4
pkg-resources==0.0.0
protobuf==3.11.3
pyaml==20.4.0
pygraphviz==1.5
pyroute2==0.5.12
python-dotenv==0.13.0
PyYAML==5.3.1
six==1.14.0
